refactor: move canvas state out of page internals
