+++ src/pages/canvas/widgets/PipelineCanvas/CanvasBackground.tsx
@@ -3,7 +3,7 @@ import { Layers as IconLayers } from 'lucide-react';
 import type React from 'react';
 import { useMemo } from 'react';
 import type { CanvasNode, ComponentGroup } from '@/entities/CanvasNode/model/types.ts';
-import { layoutNodesAtom } from './model/atoms.ts';
+import { layoutNodesAtom } from '@/features/Canvas/model/atoms';
 import { estimateNodeHeight } from './utils.ts';
 
 const CanvasBackground: React.FC = () => {
