+++ src/pages/canvas/widgets/PipelineCanvas/CanvasCodeCard.tsx
@@ -9,8 +9,8 @@ import type React from 'react';
 import { useMemo, useRef } from 'react';
 import { focusedPaneAtom } from '@/entities/AppView/model/atoms';
 import type { CanvasNode } from '@/entities/CanvasNode/model/types.ts';
+import { cardPositionsAtom, selectedNodeIdsAtom, transformAtom } from '@/features/Canvas/model/atoms';
 import CodeCard from '@/widgets/CodeCard/CodeCard.tsx';
-import { cardPositionsAtom, selectedNodeIdsAtom, transformAtom } from './model/atoms.ts';
 
 interface CanvasCodeCardProps {
   node: CanvasNode;
