+++ src/pages/canvas/widgets/PipelineCanvas/PipelineCanvas.tsx
@@ -4,12 +4,12 @@ import { useEffect, useMemo, useRef } from 'react';
 import { useHotkeys } from 'react-hotkeys-hook';
 // Atoms & Hooks
 import { filesAtom, focusedPaneAtom, fullNodeMapAtom, graphDataAtom } from '@/entities/AppView/model/atoms';
+import { openedFilesAtom, selectedNodeIdsAtom, visibleNodeIdsAtom } from '@/features/Canvas/model/atoms';
 import { symbolMetadataAtom } from '@/features/Search/UnifiedSearch/model/atoms.ts';
 import { extractSymbolMetadata } from '@/shared/symbolMetadataExtractor.ts';
 import { CanvasCodeCard } from './CanvasCodeCard.tsx';
 import CanvasConnections from './CanvasConnections.tsx';
 import D3ZoomContainer from './D3ZoomContainer.tsx';
-import { openedFilesAtom, selectedNodeIdsAtom, visibleNodeIdsAtom } from './model/atoms.ts';
 // Hooks & Sub-components
 import CopyAllCodeButton from './ui/CopyAllCodeButton';
 import ResetViewButton from './ui/ResetViewButton';
