+++ src/pages/canvas/widgets/PipelineCanvas/utils.ts
@@ -1,5 +1,4 @@
 import type { CanvasNode } from '@/entities/CanvasNode/model/types.ts';
-import type { SourceFileNode } from '@/entities/SourceFileNode/model/types.ts';
 
 // --- Constants ---
 export const LEVEL_SPACING = 850; // Horizontal space between columns
@@ -49,59 +48,4 @@ export const estimateNodeHeight = (node: CanvasNode) => {
   const totalHeight = baseHeight + visualLineCount * lineHeight;
 
   return totalHeight;
-};
-
-/**
- * Calculates which nodes are still reachable from the Entry File or Template Root.
- * Any node currently in `visibleSet` that cannot be reached is considered an "orphan" and removed.
- */
-export const pruneDetachedNodes = (
-  visibleSet: Set<string>,
-  nodeMap: Map<string, SourceFileNode>,
-  entryFile: string,
-  templateRootId: string | null
-): Set<string> => {
-  const visited = new Set<string>();
-  const queue: string[] = [];
-
-  // 1. Identify Roots within the Visible Set
-  // A node is a valid "root" source if it is in the Entry File OR it is the Template Root
-  visibleSet.forEach((id) => {
-    const node = nodeMap.get(id);
-    if (!node) return;
-
-    // Is it an anchor point?
-    // - It's the designated Template/JSX Root
-    // - OR It belongs to the entry file (e.g. setup calls, top level variables)
-    const isEntryPoint = node.filePath === entryFile;
-    const isTemplateRoot = id === templateRootId;
-
-    if (isEntryPoint || isTemplateRoot) {
-      if (!visited.has(id)) {
-        visited.add(id);
-        queue.push(id);
-      }
-    }
-  });
-
-  // 2. BFS Traversal to find all reachable nodes
-  while (queue.length > 0) {
-    const currentId = queue.shift()!;
-    const node = nodeMap.get(currentId);
-
-    if (node) {
-      node.dependencies.forEach((depId) => {
-        // We can only traverse to a node if it is currently 'visible' (enabled by user)
-        // If the user turned off a node, the path stops there.
-        if (visibleSet.has(depId) && !visited.has(depId)) {
-          visited.add(depId);
-          queue.push(depId);
-        }
-      });
-    }
-  }
-
-  // 3. The new visible set is exactly what we could reach.
-  // Anything in 'visibleSet' but NOT in 'visited' was cut off.
-  return visited;
 };
