+++ src/widgets/CodeViewer/ui/SegmentRenderer.tsx
@@ -6,9 +6,9 @@
 import { useAtomValue } from 'jotai';
 import { useMemo } from 'react';
 import { useEditorTheme } from '@/entities/AppTheme/EditorThemeProvider';
+import { visibleNodeIdsAtom } from '@/features/Canvas/model/atoms';
 import { activeLocalVariablesAtom } from '@/features/Code/FocusMode/model/atoms';
 import { LocalVariableSegment } from '@/features/Code/FocusMode/ui/LocalVariableSegment';
-import { visibleNodeIdsAtom } from '@/pages/canvas/widgets/PipelineCanvas/model/atoms';
 import type { CanvasNode } from '../../../entities/CanvasNode/model/types';
 import { buildSegmentStyle } from '../core/styler/styleBuilder';
 import type { CodeLine, CodeSegment, SegmentKind } from '../core/types/codeLine';
