+++ src/widgets/CodeViewer/ui/segments/ExternalSegment.tsx
@@ -8,12 +8,8 @@ import { useAtomValue, useSetAtom } from 'jotai';
 import type React from 'react';
 import { activeActivityPageIdAtom } from '@/app/model/activityPageAtoms';
 import { fullNodeMapAtom, hoveredIdentifierAtom } from '@/entities/AppView/model/atoms';
+import { cardPositionsAtom, transformAtom, visibleNodeIdsAtom } from '@/features/Canvas/model/atoms';
 import { useOpenFile } from '@/features/File/OpenFiles/lib/useOpenFile';
-import {
-  cardPositionsAtom,
-  transformAtom,
-  visibleNodeIdsAtom,
-} from '@/pages/canvas/widgets/PipelineCanvas/model/atoms';
 import { focusedNodeIdAtom } from '@/widgets/MainContents/IDEScrollView/model/atoms.ts';
 import type { CanvasNode } from '../../../../entities/CanvasNode/model/types';
 import type { CodeSegment, SegmentStyle } from '../../core/types';
