+++ src/widgets/JotaiDevTools/JotaiDevTools.tsx
@@ -8,13 +8,13 @@ import React, { useEffect, useState } from 'react';
 import * as themeAtoms from '@/entities/AppTheme/atoms';
 import * as appAtoms from '@/entities/AppView/model/atoms';
 import { store } from '@/entities/AppView/model/store';
+import * as canvasAtoms from '@/features/Canvas/model/atoms';
 import * as deadCodeAtoms from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/model/atoms';
 import * as codeFoldAtoms from '@/features/Code/CodeFold/model/atoms';
 import * as focusModeAtoms from '@/features/Code/FocusMode/model/atoms';
 import * as navigationAtoms from '@/features/File/Navigation/model/atoms';
 import * as filesAtoms from '@/features/File/OpenFiles/model/atoms';
 import * as searchAtoms from '@/features/Search/UnifiedSearch/model/atoms';
-import * as canvasAtoms from '@/pages/canvas/widgets/PipelineCanvas/model/atoms';
 import * as sidebarAtoms from '@/pages/explorer/widgets/ExplorerSidebar/model/atoms';
 
 // Combine all atoms for DevTools tracking
